+++ test/libsolidity/SolidityEndToEndTest.cpp
@@ -2604,6 +2604,23 @@ BOOST_AUTO_TEST_CASE(shadow_function_by_accessor)
 	BOOST_CHECK(callContractFunction("f()") == encodeArgs(2));
 }
 
+BOOST_AUTO_TEST_CASE(implement_interface_by_accessor)
+{
+	char const* sourceCode = R"(
+		contract A {
+			function f() returns (uint);
+		}
+		contract B is A {
+			uint256 public f;
+			function B() {
+				f = 2;
+			}
+		}
+	)";
+	compileAndRun(sourceCode, 0, "B");
+	BOOST_CHECK(callContractFunction("f()") == encodeArgs(2));
+}
+
 BOOST_AUTO_TEST_CASE(function_modifier_calling_functions_in_creation_context)
 {
 	char const* sourceCode = R"(